+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Grant tool pouches and crafting kits for nail upgrades
 - Respect Ascended & Radiant difficulties in godhome
 - Fix silk regeneration not working after re-entering a save
+- Clear active tools on new savefile creation
 
 ## v0.1.0